+++ src/pages/LoadUnload/components/MapLayers/MapLayers.tsx
@@ -1,17 +1,23 @@
 import { BaseLayer } from '@amsterdam/arm-core'
 
+import { AddressMarker } from '../../../../shared/components/MapLayers/AddressMarker'
 import { HighlightedFeatureLayer } from '../../../../shared/components/HighlightedFeatureLayer'
 import { topoBlackWhite } from '../../../../shared/map/mapLayers'
 
 import { useLoadUnloadMapContext } from '../../contexts/MapContext'
+import { useLoadUnloadPageContext } from '../../contexts/PageContext'
+
 import { LoadUnloadRoadSectionsLoadUnloadLayer } from './RoadSectionsLoadUnloadLayer'
 import { LoadUnloadLoadUnloadSpacesLayer } from './LoadUnloadSpacesLayer'
 
 export const LoadUnloadMapLayers = () => {
   const { detailFeature } = useLoadUnloadMapContext()
+  const { address } = useLoadUnloadPageContext()
 
   return (
     <>
+      <AddressMarker address={address} />
+
       {detailFeature.feature && (
         <HighlightedFeatureLayer
           detailFeature={{