+++ CHANGELOG.md
@@ -4,10 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/Amsterdam/bereikbaarheid-backend/compare/v2.2...HEAD)
+## [Unreleased](https://github.com/Amsterdam/bereikbaarheid-frontend/compare/v2.2.1...HEAD)
 
 ### Added
+- Load-Unload: address search
 - Load-Unload: road sections with load-unload information
+
+### Changed
+- Load-Unload: increase max zoom level to 21
+
+
+## [v2.2.1 - 2022-12-15](https://github.com/Amsterdam/bereikbaarheid-frontend/compare/v2.2...v2.2.1)
+Updated minor and patch dependencies.
 
 
 ## v2.2 - 2022-12-09